ref(dx): Move project core endpoints to core module
