+++ src/sentry/core/endpoints/project_details.py
