+++ src/sentry/core/endpoints/project_environment_details.py
