+++ src/sentry/core/endpoints/project_environments.py
