+++ src/sentry/core/endpoints/project_key_details.py
