+++ src/sentry/core/endpoints/project_team_details.py
