+++ tests/sentry/core/endpoints/test_project_details.py
