+++ tests/sentry/core/endpoints/test_project_environment_details.py
