+++ tests/sentry/core/endpoints/test_project_key_details.py
