+++ tests/sentry/core/endpoints/test_project_stats.py
