+++ tests/sentry/core/endpoints/test_project_team_details.py
